Did some overall code clean up and added better error handling
Got rid of the prefixing _ characters in some of the structs

Got rid of many of the unwraps for better error handling (or replaced them with things like unwrap_or)
In some instances now instead of panicking, it will just fail and then bring up a Windows Messagebox saying something went wrong.

Moved most of the logic from the lambda in the on_close callback to a function that gets called, from the callback.

isolated some logic around getting the FUNC_ITEMS out of the once cell into a function, to remove some of the duplicated code.
I do wonder if there is a better way to handle the FUNC_ITEMS than a static mut, but I'm not quite sure what the best way would be in a dll at the moment

Cleaned up some unnecessary indexing into the FUNC_ITEMS Vec and the bounds checking for it.

Modified the Dialog box a bit, put in some slightly more descriptive text, and set the default values for the shift to 0's